# docs/swagger.json
{
    "definitions": {
...
        "notification.Webhook": {
            "properties": {
                "active": {
                    "type": "boolean"
                },
                "createdAt": {
                    "type": "string"
                },
                "errorsCount": {
                    "type": "integer"
                },
                "lastEmitStatus": {
                    "type": "string"
                },
                "lastEmitTimestamp": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            },
            "type": "object"
        },
...
        "transports_http_endpoints_api_webhook.Request": {
            "properties": {
                "requiredAuth": {
                    "$ref": "#/definitions/transports_http_endpoints_api_webhook.RequiredAuth"
                },
                "url": {
                    "type": "string"
                }
            },
            "type": "object"
        },
        "transports_http_endpoints_api_webhook.RequiredAuth": {
            "properties": {
                "header": {
                    "type": "string"
                },
                "token": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            },
            "type": "object"
        }
    },
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/webhook": {
            "delete": {
                "consumes": [
                    "*/*"
                ],
                "parameters": [
                    {
                        "description": "URL of webhook to revoke",
                        "in": "query",
                        "name": "url",
                        "required": true,
                        "type": "string"
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "summary": "Revoke webhook",
                "tags": [
                    "webhooks"
                ]
            },
            "get": {
                "consumes": [
                    "*/*"
                ],
                "parameters": [
                    {
                        "description": "URL of webhook to check",
                        "in": "query",
                        "name": "url",
                        "required": true,
                        "type": "string"
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/notification.Webhook"
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "summary": "Get webhook",
                "tags": [
                    "webhooks"
                ]
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "description": "Webhook to register",
                        "in": "body",
                        "name": "data",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/transports_http_endpoints_api_webhook.Request"
                        }
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/notification.Webhook"
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "summary": "Register new webhook",
                "tags": [
                    "webhooks"
                ]
            }
        }
    },
    "securityDefinitions": {
        "Bearer": {
            "in": "header",
            "name": "Authorization",
            "type": "apiKey"
        }
    },
    "swagger": "2.0"
}